removed clock slot

+++ contracts/clearing-house/src/helpers/fees.rs
@@ -129,7 +129,7 @@ fn calculate_referral_reward_and_referee_discount(
 
 pub fn calculate_order_fee_tier(
     fee_structure: &FeeStructure,
-    discount_token_amt: u64,
+    discount_token_amt: u128,
 ) -> Result<OrderDiscountTier, ContractError> {
     if discount_token_amt == 0 {
         return Ok(OrderDiscountTier::None);
@@ -171,8 +171,8 @@ pub fn calculate_fee_for_order(
     fee_structure: &FeeStructure,
     filler_reward_structure: &OrderFillerRewardStructure,
     order_fee_tier: &OrderDiscountTier,
-    order_ts: i64,
-    now: i64,
+    order_ts: u64,
+    now: u64,
     referrer: &Option<Addr>,
     filler_is_user: bool,
     quote_asset_amount_surplus: u128,
@@ -258,8 +258,8 @@ fn calculate_token_discount_for_limit_order(
 
 fn calculate_filler_reward(
     fee: u128,
-    order_ts: i64,
-    now: i64,
+    order_ts: u64,
+    now: u64,
     filler_reward_structure: &OrderFillerRewardStructure,
 ) -> Result<u128, ContractError> {
     // incentivize keepers to prioritize filling older orders (rather than just largest orders)
